Move Role and Type code to a separate folder
