+++ api/user.js
@@ -6,7 +6,7 @@ const express = require("express");
 const url = require("url");
 const validator = require("validator");
 
-const {Role} = require("../utils/role");
+const {Role} = require("../entities/role");
 const userModel = require("../models/user");
 const {
   userSchema,
