+++ entities/type.js
