+++ utils/auth.js
@@ -7,7 +7,7 @@ const jwt = require("jsonwebtoken");
 const url = require("url");
 const xml2js = require("xml2js");
 
-const {Role} = require("./role");
+const {Role} = require("../entities/role");
 
 const JWT_SECRET_KEY = process.env.JWT_SECRET_KEY;
